Things/Path_Network/requirements.txt
numpy>=1.21.0
networkx>=2.6.0
scipy>=1.7.0
matplotlib>=3.4.0
torch>=1.9.0
pandas>=1.3.0
seaborn>=0.11.0
pytest>=6.2.0
black>=21.5b2
mypy>=0.910
pylint>=2.8.0
pyyaml>=5.4.1
plotly>=5.3.0
imageio>=2.9.0
tqdm>=4.62.0
scikit-learn>=0.24.0
ipywidgets>=7.6.0
moviepy>=1.0.3
mayavi>=4.7.0 